refactor(deliverable-template): Rams指摘2点修正 — 500行超テスト分割 + §確定C _from_row修正
- test_crud.py (613行 / 545行) を test_crud/ ディレクトリに分割
  - DTR: test_protocol_and_crud.py (TC-IT-DTR-001~009)
         test_schema_and_conversions.py (TC-IT-DTR-010~019)
  - RPR: test_crud_basic.py (TC-IT-RPR-001~009)
         test_crud_constraints.py (TC-IT-RPR-010~015)
- _from_row を DeliverableTemplate.model_validate / RoleProfile.model_validate 経由に修正
  (§確定 C: post-validator で不変条件を再実行する)
